Refactor QuantizedVectorsReader to be able to consolidate the PI for lucene 104 quantize values
