+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene99/TestLucene99HnswQuantizedVectorsFormat.java
@@ -52,7 +52,7 @@ import org.apache.lucene.tests.index.BaseKnnVectorsFormatTestCase;
 import org.apache.lucene.tests.util.TestUtil;
 import org.apache.lucene.util.SameThreadExecutorService;
 import org.apache.lucene.util.VectorUtil;
-import org.apache.lucene.util.quantization.QuantizedByteVectorValues;
+import org.apache.lucene.util.quantization.LegacyQuantizedByteVectorValues;
 import org.apache.lucene.util.quantization.ScalarQuantizer;
 import org.junit.Before;
 
@@ -277,8 +277,8 @@ public class TestLucene99HnswQuantizedVectorsFormat extends BaseKnnVectorsFormat
           knnVectorsReader = knnVectorsReader.unwrapReaderForField("f");
           if (knnVectorsReader instanceof Lucene99HnswVectorsReader hnswReader) {
             assertNotNull(hnswReader.getQuantizationState("f"));
-            QuantizedByteVectorValues quantizedByteVectorValues =
-                hnswReader.getQuantizedVectorValues("f");
+            LegacyQuantizedByteVectorValues quantizedByteVectorValues =
+                (LegacyQuantizedByteVectorValues) hnswReader.getQuantizedVectorValues("f");
             for (int ord = 0; ord < quantizedByteVectorValues.size(); ord++) {
               byte[] vector = quantizedByteVectorValues.vectorValue(ord);
               float offset = quantizedByteVectorValues.getScoreCorrectionConstant(ord);
