+++ lucene/core/src/java25/org/apache/lucene/internal/vectorization/Lucene99MemorySegmentScalarQuantizedVectorScorer.java
@@ -35,7 +35,7 @@ import org.apache.lucene.util.VectorUtil;
 import org.apache.lucene.util.hnsw.RandomVectorScorer;
 import org.apache.lucene.util.hnsw.RandomVectorScorerSupplier;
 import org.apache.lucene.util.hnsw.UpdateableRandomVectorScorer;
-import org.apache.lucene.util.quantization.QuantizedByteVectorValues;
+import org.apache.lucene.util.quantization.LegacyQuantizedByteVectorValues;
 import org.apache.lucene.util.quantization.ScalarQuantizer;
 
 class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsScorer {
@@ -51,7 +51,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
   public RandomVectorScorerSupplier getRandomVectorScorerSupplier(
       VectorSimilarityFunction similarityFunction, KnnVectorValues vectorValues)
       throws IOException {
-    if (vectorValues instanceof QuantizedByteVectorValues quantized
+    if (vectorValues instanceof LegacyQuantizedByteVectorValues quantized
         && quantized.getSlice() instanceof MemorySegmentAccessInput input) {
       return new RandomVectorScorerSupplierImpl(similarityFunction, quantized, input);
     }
@@ -62,7 +62,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
   public RandomVectorScorer getRandomVectorScorer(
       VectorSimilarityFunction similarityFunction, KnnVectorValues vectorValues, float[] target)
       throws IOException {
-    if (vectorValues instanceof QuantizedByteVectorValues quantized
+    if (vectorValues instanceof LegacyQuantizedByteVectorValues quantized
         && quantized.getSlice() instanceof MemorySegmentAccessInput input) {
       if (Constants.NATIVE_DOT_PRODUCT_ENABLED) {
         return new NativeRandomVectorScorerImpl(similarityFunction, quantized, input, target);
@@ -98,7 +98,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
 
     RandomVectorScorerBase(
         VectorSimilarityFunction similarityFunction,
-        QuantizedByteVectorValues values,
+        LegacyQuantizedByteVectorValues values,
         MemorySegmentAccessInput input) {
       super(values);
 
@@ -208,7 +208,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
 
     RandomVectorScorerImpl(
         VectorSimilarityFunction similarityFunction,
-        QuantizedByteVectorValues values,
+        LegacyQuantizedByteVectorValues values,
         MemorySegmentAccessInput input,
         float[] target) {
       super(similarityFunction, values, input);
@@ -255,7 +255,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
   private static class NativeRandomVectorScorerImpl extends RandomVectorScorerImpl {
     NativeRandomVectorScorerImpl(
         VectorSimilarityFunction similarityFunction,
-        QuantizedByteVectorValues values,
+        LegacyQuantizedByteVectorValues values,
         MemorySegmentAccessInput input,
         float[] target) {
       super(similarityFunction, values, input, target);
@@ -294,7 +294,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
 
   private record RandomVectorScorerSupplierImpl(
       VectorSimilarityFunction similarityFunction,
-      QuantizedByteVectorValues values,
+      LegacyQuantizedByteVectorValues values,
       MemorySegmentAccessInput input)
       implements RandomVectorScorerSupplier {
 
@@ -319,7 +319,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
 
     UpdateableRandomVectorScorerImpl(
         VectorSimilarityFunction similarityFunction,
-        QuantizedByteVectorValues values,
+        LegacyQuantizedByteVectorValues values,
         MemorySegmentAccessInput input) {
       super(similarityFunction, values, input);
     }
@@ -373,7 +373,7 @@ class Lucene99MemorySegmentScalarQuantizedVectorScorer implements FlatVectorsSco
 
     NativeUpdateableRandomVectorScorerImpl(
         VectorSimilarityFunction similarityFunction,
-        QuantizedByteVectorValues values,
+        LegacyQuantizedByteVectorValues values,
         MemorySegmentAccessInput input) {
       super(similarityFunction, values, input);
     }
