+++ lucene/test-framework/src/java/org/apache/lucene/tests/index/RandomCodec.java
@@ -34,7 +34,6 @@ import org.apache.lucene.codecs.PointsWriter;
 import org.apache.lucene.codecs.PostingsFormat;
 import org.apache.lucene.codecs.blocktreeords.BlockTreeOrdsPostingsFormat;
 import org.apache.lucene.codecs.lucene104.Lucene104HnswScalarQuantizedVectorsFormat;
-import org.apache.lucene.codecs.lucene104.Lucene104ScalarQuantizedVectorsFormat;
 import org.apache.lucene.codecs.lucene90.Lucene90DocValuesFormat;
 import org.apache.lucene.codecs.lucene90.Lucene90PointsReader;
 import org.apache.lucene.codecs.lucene90.Lucene90PointsWriter;
@@ -62,6 +61,7 @@ import org.apache.lucene.tests.util.TestUtil;
 import org.apache.lucene.util.IORunnable;
 import org.apache.lucene.util.bkd.BKDConfig;
 import org.apache.lucene.util.bkd.BKDWriter;
+import org.apache.lucene.util.quantization.QuantizedByteVectorValues;
 
 /**
  * Codec that assigns per-field random postings formats.
@@ -275,21 +275,21 @@ public class RandomCodec extends AssertingCodec {
             concurrentKnnMerging ? ForkJoinPool.commonPool() : null,
             0),
         new Lucene104HnswScalarQuantizedVectorsFormat(
-            Lucene104ScalarQuantizedVectorsFormat.ScalarEncoding.SEVEN_BIT,
+            QuantizedByteVectorValues.ScalarEncoding.SEVEN_BIT,
             TestUtil.nextInt(random, 5, 50),
             TestUtil.nextInt(random, 10, 50),
             concurrentKnnMerging ? TestUtil.nextInt(random, 2, 8) : 1,
             concurrentKnnMerging ? ForkJoinPool.commonPool() : null,
             0),
         new Lucene104HnswScalarQuantizedVectorsFormat(
-            Lucene104ScalarQuantizedVectorsFormat.ScalarEncoding.UNSIGNED_BYTE,
+            QuantizedByteVectorValues.ScalarEncoding.UNSIGNED_BYTE,
             TestUtil.nextInt(random, 5, 50),
             TestUtil.nextInt(random, 10, 50),
             concurrentKnnMerging ? TestUtil.nextInt(random, 2, 8) : 1,
             concurrentKnnMerging ? ForkJoinPool.commonPool() : null,
             0),
         new Lucene104HnswScalarQuantizedVectorsFormat(
-            Lucene104ScalarQuantizedVectorsFormat.ScalarEncoding.PACKED_NIBBLE,
+            QuantizedByteVectorValues.ScalarEncoding.PACKED_NIBBLE,
             TestUtil.nextInt(random, 5, 50),
             TestUtil.nextInt(random, 10, 50),
             concurrentKnnMerging ? TestUtil.nextInt(random, 2, 8) : 1,
